@@ -23,17 +23,20 @@
  */
 
 apply plugin: 'com.android.application'
+apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "org.angelmariages.rodalieswidget"
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 22
+        targetSdkVersion 28
+        multiDexEnabled true
         versionCode 37
         versionName "1.8"
         vectorDrawables.useSupportLibrary = true
+        resValue "string", "preferences_provider_authority", "${applicationId}.preferencesprovider"
     }
     buildTypes {
         release {
@@ -52,41 +55,34 @@ android {
     }
 }
 
-buildscript {
-    repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.22.1'
-    }
-}
-
-apply plugin: 'io.fabric'
-
-repositories {
-    maven { url 'https://maven.fabric.io/public' }
-}
-
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.google.code.gson:gson:2.8.2'
-    implementation 'com.takisoft.fix:preference-v7:27.1.0.0'
-    implementation 'com.takisoft.fix:preference-v7-ringtone:27.1.0.0'
-    implementation 'com.android.support:appcompat-v7:27.1.0'
-    implementation 'com.android.support:design:27.1.0'
-    implementation 'com.google.firebase:firebase-database:12.0.1'
-    implementation 'com.google.firebase:firebase-config:12.0.1'
-    implementation 'com.google.firebase:firebase-messaging:12.0.1'
-    implementation 'com.google.firebase:firebase-firestore:12.0.1'
-    implementation 'com.google.firebase:firebase-ads:12.0.1'
-    implementation 'com.android.support:support-v4:27.1.0'
-    implementation 'com.android.support:support-vector-drawable:27.1.0'
+
+    implementation 'com.takisoft.preferencex:preferencex:1.0.0'
+    implementation 'com.takisoft.preferencex:preferencex-ringtone:1.0.0'
+
+    implementation 'com.google.android.material:material:1.1.0-alpha01'
+
+    implementation 'androidx.browser:browser:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.0.1'
+
+    implementation 'androidx.multidex:multidex:2.0.0'
+
+    implementation 'com.google.firebase:firebase-core:16.0.5'
+    implementation 'com.google.firebase:firebase-database:16.0.5'
+    implementation 'com.google.firebase:firebase-config:16.1.0'
+    implementation 'com.google.firebase:firebase-messaging:17.3.4'
+    implementation 'com.google.firebase:firebase-firestore:17.1.3'
+    implementation 'com.google.firebase:firebase-ads:17.1.1'
+
     implementation 'com.github.kizitonwose.colorpreference:support:1.0.6'
 
-    implementation 'net.grandcentrix.tray:tray:0.12.0'
+    implementation 'com.mengdd.preferencesprovider:preferences-provider:0.2.0'
 
-    implementation('com.crashlytics.sdk.android:crashlytics:2.8.0@aar') {
+    implementation('com.crashlytics.sdk.android:crashlytics:2.9.6') {
         transitive = true
     }
 }
