@@ -24,7 +24,7 @@
 
 package org.angelmariages.rodalieswidget;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
