@@ -27,13 +27,13 @@ package org.angelmariages.rodalieswidget;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.Nullable;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -50,7 +50,7 @@ import com.google.firebase.database.ValueEventListener;
 
 import org.angelmariages.rodalieswidget.utils.U;
 
-public class FirstTimeActivity extends AppCompatActivity {
+public class TutorialActivity extends AppCompatActivity {
 	private ViewPager viewPager;
 	private LinearLayout dotsLayout;
 	private Button btnSkip;
@@ -68,17 +68,15 @@ public class FirstTimeActivity extends AppCompatActivity {
 		}
 
 		// Making notification bar transparent
-		if (Build.VERSION.SDK_INT >= 21) {
-			getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_STABLE | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
-		}
+		getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_STABLE | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
 		setContentView(R.layout.first_time_activity);
 
 		changeStatusBarColor();
 
-		viewPager = (ViewPager) findViewById(R.id.view_pager);
-		dotsLayout = (LinearLayout) findViewById(R.id.layoutDots);
-		btnSkip = (Button) findViewById(R.id.skipButton);
-		btnNext = (Button) findViewById(R.id.nextButton);
+		viewPager = findViewById(R.id.view_pager);
+		dotsLayout = findViewById(R.id.layoutDots);
+		btnSkip = findViewById(R.id.skipButton);
+		btnNext = findViewById(R.id.nextButton);
 
 		layouts = new int[]{
 				R.layout.tutorial_slide_1,
@@ -96,7 +94,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 				DatabaseReference tutorialReference = U.getFirebaseDatabase().getReference("tutorial");
 				tutorialReference.addListenerForSingleValueEvent(new ValueEventListener() {
 					@Override
-					public void onDataChange(DataSnapshot dataSnapshot) {
+					public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 						DataSnapshot skip = dataSnapshot.child("skip");
 
 						if (skip.exists()) {
@@ -108,7 +106,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 					}
 
 					@Override
-					public void onCancelled(DatabaseError databaseError) {
+					public void onCancelled(@NonNull DatabaseError databaseError) {
 						U.log("FirebaseError (skip): " + databaseError.getMessage());
 					}
 				});
@@ -129,7 +127,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 					DatabaseReference tutorialReference = U.getFirebaseDatabase().getReference("tutorial");
 					tutorialReference.addListenerForSingleValueEvent(new ValueEventListener() {
 						@Override
-						public void onDataChange(DataSnapshot dataSnapshot) {
+						public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 							DataSnapshot skip = dataSnapshot.child("start");
 
 							if (skip.exists()) {
@@ -141,7 +139,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 						}
 
 						@Override
-						public void onCancelled(DatabaseError databaseError) {
+						public void onCancelled(@NonNull DatabaseError databaseError) {
 							U.log("FirebaseError (skip): " + databaseError.getMessage());
 						}
 					});
@@ -151,7 +149,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 	}
 
 	private void startSettings() {
-		startActivity(new Intent(FirstTimeActivity.this, MainActivity.class));
+		startActivity(new Intent(TutorialActivity.this, MainActivity.class));
 
 		PreferenceManager.getDefaultSharedPreferences(this).edit().putBoolean("tutorial_viewed", true).apply();
 
@@ -161,10 +159,12 @@ public class FirstTimeActivity extends AppCompatActivity {
 	private class MyViewPagerAdapter extends PagerAdapter {
 		private LayoutInflater layoutInflater;
 
-		MyViewPagerAdapter() { }
+		MyViewPagerAdapter() {
+		}
 
+		@NonNull
 		@Override
-		public Object instantiateItem(ViewGroup container, int position) {
+		public Object instantiateItem(@NonNull ViewGroup container, int position) {
 			layoutInflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
 
 			View view = layoutInflater.inflate(layouts[position], container, false);
@@ -179,13 +179,13 @@ public class FirstTimeActivity extends AppCompatActivity {
 		}
 
 		@Override
-		public boolean isViewFromObject(View view, Object obj) {
+		public boolean isViewFromObject(@NonNull View view, @NonNull Object obj) {
 			return view == obj;
 		}
 
 
 		@Override
-		public void destroyItem(ViewGroup container, int position, Object object) {
+		public void destroyItem(@NonNull ViewGroup container, int position, @NonNull Object object) {
 			View view = (View) object;
 			container.removeView(view);
 		}
@@ -209,7 +209,7 @@ public class FirstTimeActivity extends AppCompatActivity {
 			DatabaseReference tutorialReference = U.getFirebaseDatabase().getReference("tutorial");
 			tutorialReference.addListenerForSingleValueEvent(new ValueEventListener() {
 				@Override
-				public void onDataChange(DataSnapshot dataSnapshot) {
+				public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 					DataSnapshot skip = dataSnapshot.child("tutorial_" + position);
 
 					if (skip.exists()) {
@@ -221,17 +221,19 @@ public class FirstTimeActivity extends AppCompatActivity {
 				}
 
 				@Override
-				public void onCancelled(DatabaseError databaseError) {
+				public void onCancelled(@NonNull DatabaseError databaseError) {
 					U.log("FirebaseError (skip): " + databaseError.getMessage());
 				}
 			});
 		}
 
 		@Override
-		public void onPageScrolled(int arg0, float arg1, int arg2) { }
+		public void onPageScrolled(int arg0, float arg1, int arg2) {
+		}
 
 		@Override
-		public void onPageScrollStateChanged(int arg0) { }
+		public void onPageScrollStateChanged(int arg0) {
+		}
 	};
 
 	private void addBottomDots(int currentPage) {
@@ -255,10 +257,8 @@ public class FirstTimeActivity extends AppCompatActivity {
 	}
 
 	private void changeStatusBarColor() {
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-			Window window = getWindow();
-			window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-			window.setStatusBarColor(Color.TRANSPARENT);
-		}
+		Window window = getWindow();
+		window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
+		window.setStatusBarColor(Color.TRANSPARENT);
 	}
 }
